requirements.txt
Flask==3.0.1
git+https://github.com/huggingface/transformers.git
# transformers==4.37.1
Pillow==10.2.0
torch==2.1.2